fix: non-symbolic enum members cannot be rendered
If an starts with a non-alpha character, we now prefix the member name with `VALUE_` so it can be used as a symbol.
